third_party/WebKit/LayoutTests/platform/win/external/wpt/shadow-dom/untriaged/elements-and-dom-objects/extensions-to-event-interface/event-path-001-expected.txt

This is a testharness.js-based test.
FAIL event.composedPath() cross the shadow boundary assert_equals: composedPath().length expected 7 but got 2
Harness: the test ran to completion.
